requirements.txt
attrs==21.4.0
click==8.0.4
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Cors==3.0.10
flask-expects-json==1.7.0
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
mongoengine==0.23.1
pymongo==3.12.3
pyrsistent==0.18.1
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
